requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.15
python-multipart==0.0.20
redis==5.2.1
pydantic==2.10.5
python-dotenv==1.0.1
langchain-groq==0.2.3